splitting *.h to *.hpp and *.h
