+++ examples/BMA020SoftI2C/BMA020SoftI2C.ino
@@ -26,7 +26,7 @@
 #define I2C_FASTMODE 1
 #endif
 
-#include <SoftI2CMaster.h>
+#include <SoftI2CMaster.hpp>
 #include <avr/io.h>
