+++ examples/I2CShell/I2CShell.ino
@@ -82,7 +82,7 @@
 #define I2C_FASTMODE 1
 #endif
 
-#include <SoftI2CMaster.h>
+#include <SoftI2CMaster.hpp>
 
 // constants for this program
 #define LINELEN 80
