+++ examples/TSL2561Soft/TSL2561Soft.ino
@@ -17,7 +17,7 @@
 #define I2C_FASTMODE 1
 #endif
 
-#include <SoftI2CMaster.h>
+#include <SoftI2CMaster.hpp>
 #include "TSL2561Soft.h"
 
 #define ADDR 0x72
