+++ tests/softi2c/softi2c.ino
@@ -25,7 +25,7 @@
 #define I2C_FASTMODE 1
 #endif
 
-#include <SoftI2CMaster.h>
+#include <SoftI2CMaster.hpp>
 
 void setup() {
   pinMode(LEDPIN, OUTPUT);
